pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "moka-news"
version = "0.1.0"
description = "Morning Persona News - A TUI RSS news aggregator with AI summaries"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Gianluigi Calcaterra"}
]
dependencies = [
    "feedparser>=6.0.10",
    "textual>=0.40.0",
    "openai>=1.0.0",
    "anthropic>=0.7.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
moka-news = "moka_news.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["moka_news*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.ruff]
line-length = 88
target-version = "py38"
